+++ CHANGELOG.md
@@ -3,7 +3,41 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
-## [20.07.0] - 2020-08-10
+## [2.2007.2] - 2020-08-31
+
+### Fixed
+  - Compaction: Use separate compactors for L0, L1 (#1466)
+  - Rework Block and Index cache (#1473)
+  - Add IsClosed method (#1478)
+  - Cleanup: Avoid truncating in vlog.Open on error (#1465)
+  - Cleanup: Do not close cache before compactions (#1464)
+
+### New APIs
+- Badger.DB
+  - BlockCacheMetrics (#1473)
+  - IndexCacheMetrics (#1473)
+- Badger.Option
+  - WithBlockCacheSize (#1473)
+  - WithIndexCacheSize (#1473)
+
+### Removed APIs [Breaking Changes]
+- Badger.DB
+  - DataCacheMetrics (#1473)
+  - BfCacheMetrics (#1473)
+- Badger.Option
+  - WithMaxCacheSize (#1473)
+  - WithMaxBfCacheSize (#1473)
+  - WithKeepBlockIndicesInCache (#1473)
+  - WithKeepBlocksInCache (#1473)
+
+## [2.2007.1] - 2020-08-19
+
+### Fixed
+  - Remove vlog file if bootstrap, syncDir or mmap fails (#1434)
+  - levels: Compaction incorrectly drops some delete markers (#1422)
+  - Replay: Update head for LSM entires also (#1456)
+
+## [2.2007.0] - 2020-08-10
 
 ### Fixed
   - Add a limit to the size of the batches sent over a stream. (#1412)
@@ -403,8 +437,10 @@ Bug fix:
 ## [1.0.1] - 2017-11-06
 * Fix an uint16 overflow when resizing key slice
 
-[Unreleased]: https://github.com/dgraph-io/badger/compare/v20.07.0...HEAD
-[20.07.0]: https://github.com/dgraph-io/badger/compare/v2.0.3...v20.07.0
+[Unreleased]: https://github.com/dgraph-io/badger/compare/v2.2007.2...HEAD
+[2.2007.2]: https://github.com/dgraph-io/badger/compare/v2.2007.1...v2.2007.2
+[2.2007.1]: https://github.com/dgraph-io/badger/compare/v2.2007.0...v2.2007.1
+[2.2007.0]: https://github.com/dgraph-io/badger/compare/v2.0.3...v2.2007.0
 [2.0.3]: https://github.com/dgraph-io/badger/compare/v2.0.2...v2.0.3
 [2.0.2]: https://github.com/dgraph-io/badger/compare/v2.0.1...v2.0.2
 [2.0.1]: https://github.com/dgraph-io/badger/compare/v2.0.0...v2.0.1